Fixing update to SIP

+++ ste/sourceInfoProvider-Local_linux-arm.go
@@ -117,8 +117,8 @@ func (s statTAdapter) BTime() time.Time {
 	return time.Time{}
 }*/
 
-func (s statTAdapter) NLink() uint32 {
-	return s.Nlink
+func (s statTAdapter) NLink() uint64 {
+	return uint64(s.Nlink)
 }
 
 /*func (s statTAdapter) Owner() uint32 {
